# docs/rest/REST-API-OPENAPI.yaml
openapi: 3.1.0
info:
  title: eMuleBB REST API (canonical superset, emulebb-rust)
  version: 1.0.0
  x-contract-version: "1.0.0"
  x-contract-baseline: "eMuleBB 0.7.3 contract"
  summary: The one capability-gated /api/v1 contract; emulebb-rust owns the superset.
  description: >
    Source-of-truth for the single capability-gated `/api/v1` contract. There is
    one contract; implementations differ by the capabilities they advertise, not
    by a separate spec. emulebb-rust is the superset (it leads the contract
    version and adds indexing/Arr capabilities); the frozen eMuleBB MFC client is
    a subset advertising fewer capabilities. It is baselined on the eMuleBB
    `0.7.3` contract and evolves on its own independent **contract version** line
    (`x-contract-version`, semver), decoupled from any product release tag.
    Controllers (TrackMuleBB) read `GET /api/v1/capabilities`, target a contract
    version range, and call only advertised operations.
    Versioning policy: additive endpoints/fields bump the minor; breaking changes
    bump the major with a documented migration. Capability model:
    `emulebb-tooling/docs/active/API-V1-COMPATIBILITY.md`.
    The API uses `X-API-Key` authentication only; adapter compatibility
    (`/api/v2` qBit-compat, Torznab) must not weaken the native route shape.
servers:
  - url: http://{host}:{port}/api/v1
    variables:
      host:
        default: 127.0.0.1
      port:
        default: "4711"
...
    variables:
      host:
...
security:
  - ApiKeyAuth: []
...
paths:
  /app:
    get:
      tags: [App]
      summary: Return application identity, runtime state, and static capabilities.
      operationId: getApp
      responses:
        "200":
          $ref: "#/components/responses/AppResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /app/preferences:
    get:
      tags: [App]
      summary: Return controller-relevant preferences.
      operationId: getPreferences
      responses:
        "200":
          $ref: "#/components/responses/PreferencesResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [App]
      summary: Update controller-relevant preferences.
      operationId: patchPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesPatch"
      responses:
        "200":
          $ref: "#/components/responses/PreferencesResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /app/shutdown:
    post:
      tags: [App]
      summary: Request clean eMule process shutdown.
      operationId: shutdownApp
      description: >
        This endpoint closes eMule only. Operating-system shutdown and reboot
        actions from the legacy WebServer are intentionally not part of the
        REST API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShutdownRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkAcceptedResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /status:
    get:
      tags: [Stats]
      summary: Return current network, transfer, server, and Kad status.
      operationId: getStatus
      responses:
        "200":
          $ref: "#/components/responses/StatusResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /snapshot:
    get:
      tags: [Stats]
      summary: Return a controller dashboard snapshot.
      operationId: getSnapshot
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/SnapshotResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /stats:
    get:
      tags: [Stats]
      summary: Return detailed cumulative and session statistics.
      operationId: getStats
      responses:
        "200":
          $ref: "#/components/responses/StatsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
...
  responses:
    AppResponse:
      description: Application envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppEnvelope"
...
    ErrorResponse:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    MethodNotAllowedResponse:
      description: Error envelope for an unsupported HTTP method on a known route.
      headers:
        Allow:
          description: Comma-separated list of HTTP methods supported by this route.
          required: true
          schema:
            type: string
          example: "GET, POST"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    OkAcceptedResponse:
      description: Accepted operation envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OkEnvelope"
...
    PreferencesResponse:
      description: Preferences envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PreferencesEnvelope"
...
    SnapshotResponse:
      description: Snapshot envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SnapshotEnvelope"
    StatsResponse:
      description: Stats envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StatsEnvelope"
    StatusResponse:
      description: Status envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StatusEnvelope"
...
  schemas:
    EnvelopeMeta:
      type: object
      additionalProperties: false
      required: [apiVersion]
      properties:
        apiVersion:
          type: string
        generatedAt:
          type: string
          format: date-time
    Ok:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
    OkEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Ok"
        meta:
          $ref: "#/components/schemas/EnvelopeMeta"
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code:
              type: string
              description: Stable machine-readable native REST error code.
              enum:
                - INVALID_ARGUMENT
                - UNAUTHORIZED
                - METHOD_NOT_ALLOWED
                - NOT_FOUND
                - INVALID_STATE
                - SERVICE_BUSY
                - EMULE_UNAVAILABLE
                - EMULE_ERROR
            message:
              type: string
              description: Human-readable error summary.
            details:
              type: object
              description: >-
                Bounded structured diagnostic fields, empty when no details are
                available. For INVALID_ARGUMENT failures on a bounded scalar
                parameter (such as an out-of-range limit or offset), details
                carries the offending "field" name and a "constraint" string of
                the form "min..max" (for example {"field": "limit", "constraint":
                "1..1000"}).
              additionalProperties: true
              properties:
                field:
                  type: string
                  description: Name of the request parameter that failed validation.
                constraint:
                  type: string
                  description: Inclusive numeric bounds the field must satisfy, formatted as "min..max".
    AppEnvelope:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/App"
...
    EnvelopeBase:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/EnvelopeMeta"
...
    App:
      type: object
      additionalProperties: false
      required: [name, version, apiVersion, lifecycle, capabilities]
      properties:
        name:
          type: string
        version:
          type: string
        apiVersion:
          type: string
        build:
          type: string
        platform:
          type: string
        lifecycle:
          $ref: "#/components/schemas/AppLifecycle"
        elevated:
          type: boolean
        webServer:
          type: object
          additionalProperties: false
          properties:
            https:
              type: boolean
            allowedIps:
              type: array
              items:
                type: string
        capabilities:
          type: object
          additionalProperties:
            type: boolean
...
    PreferencesEnvelope:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Preferences"
    Preferences:
      type: object
      additionalProperties: false
      properties:
        downloadLimitKiBps:
          type: integer
          minimum: 1
          maximum: 4294967294
        uploadLimitKiBps:
          type: integer
          minimum: 1
          maximum: 4294967294
        maxSourcesPerFile:
          type: integer
          minimum: 1
          maximum: 2147483647
        maxConnections:
          type: integer
          minimum: 1
          maximum: 2147483647
        maxConnectionsPerFiveSeconds:
          type: integer
          minimum: 1
          maximum: 2147483647
        uploadClientDataRate:
          type: integer
          minimum: 1
          maximum: 4294967295
        maxUploadSlots:
          type: integer
          minimum: 1
          maximum: 64
        uploadSlotElasticPercent:
          type: integer
          minimum: 0
          maximum: 100
        queueSize:
          type: integer
          minimum: 2000
          maximum: 10000
        autoConnect:
          type: boolean
        newAutoUp:
          type: boolean
        newAutoDown:
          type: boolean
        creditSystem:
          type: boolean
        safeServerConnect:
          type: boolean
        networkKademlia:
          type: boolean
        networkEd2k:
          type: boolean
        downloadAutoBroadbandIo:
          type: boolean
          description: Enables adaptive broadband download write buffering based on currently available RAM.
    PreferencesPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        downloadLimitKiBps:
          type: integer
          minimum: 1
          maximum: 4294967294
        uploadLimitKiBps:
          type: integer
          minimum: 1
          maximum: 4294967294
        maxSourcesPerFile:
          type: integer
          minimum: 1
          maximum: 2147483647
        maxConnections:
          type: integer
          minimum: 1
          maximum: 2147483647
        maxConnectionsPerFiveSeconds:
          type: integer
          minimum: 1
          maximum: 2147483647
        uploadClientDataRate:
          type: integer
          minimum: 1
          maximum: 4294967295
        maxUploadSlots:
          type: integer
          minimum: 1
          maximum: 64
        uploadSlotElasticPercent:
          type: integer
          minimum: 0
          maximum: 100
        queueSize:
          type: integer
          minimum: 2000
          maximum: 10000
        autoConnect:
          type: boolean
        newAutoUp:
          type: boolean
        newAutoDown:
          type: boolean
        creditSystem:
          type: boolean
        safeServerConnect:
          type: boolean
        networkKademlia:
          type: boolean
        networkEd2k:
          type: boolean
        downloadAutoBroadbandIo:
          type: boolean
          description: Enables adaptive broadband download write buffering based on currently available RAM.
    StatusEnvelope:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Status"
    Status:
      type: object
      additionalProperties: false
      required: [lifecycle, stats, servers, kad, network, sharedStartupCache, runtimeDiagnostics]
      properties:
        lifecycle:
          $ref: "#/components/schemas/AppLifecycle"
        servers:
          $ref: "#/components/schemas/ServerStatus"
        kad:
          $ref: "#/components/schemas/Kad"
        network:
          $ref: "#/components/schemas/NetworkStatus"
        stats:
          $ref: "#/components/schemas/Stats"
        sharedStartupCache:
          $ref: "#/components/schemas/SharedStartupCache"
        runtimeDiagnostics:
          $ref: "#/components/schemas/RuntimeDiagnostics"
    StatsEnvelope:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Stats"
...
    NetworkStatus:
      type: object
      additionalProperties: false
      required: [ports, binding, vpnGuard]
      properties:
        ports:
          type: object
          additionalProperties: false
          required: [tcp, udp, serverUdp]
          properties:
            tcp:
              type: integer
              minimum: 0
              maximum: 65535
            udp:
              type: integer
              minimum: 0
              maximum: 65535
            serverUdp:
              type: integer
              minimum: 0
              maximum: 65535
        binding:
          type: object
          additionalProperties: false
          required: [configuredAddress, configuredInterfaceId, configuredInterfaceName, activeConfiguredAddress, activeInterfaceId, activeInterfaceName, activeInterfaceIndex, resolveResult]
          properties:
            configuredAddress:
              type: string
            configuredInterfaceId:
              type: string
            configuredInterfaceName:
              type: string
            activeConfiguredAddress:
              type: string
            activeInterfaceId:
              type: string
            activeInterfaceName:
              type: string
            activeInterfaceIndex:
              type: integer
              minimum: 0
            resolveResult:
              type: string
              description: Stable bind-resolution outcome token.
              enum: [default, resolved, interfacenotfound, interfacenameambiguous, interfacehasnoaddress, addressnotfoundoninterface, addressnotfound, unknown]
        vpnGuard:
          type: object
          additionalProperties: false
          required: [enabled, mode, allowedPublicIpCidrs, startupBlocked, startupBlockReason]
          properties:
            enabled:
              type: boolean
            mode:
              type: string
              enum: ["off", block]
            allowedPublicIpCidrs:
              type: string
            startupBlocked:
              type: boolean
            startupBlockReason:
              type: string
...
    SnapshotEnvelope:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              additionalProperties: false
              required:
                - app
                - status
                - transfers
                - sharedFiles
                - uploads
                - uploadQueue
                - servers
                - kad
                - network
                - logs
              properties:
                app:
                  $ref: "#/components/schemas/App"
                status:
                  $ref: "#/components/schemas/Status"
                transfers:
                  type: array
                  items:
                    $ref: "#/components/schemas/Transfer"
                sharedFiles:
                  type: array
                  items:
                    $ref: "#/components/schemas/SharedFile"
                uploads:
                  type: array
                  items:
                    $ref: "#/components/schemas/Upload"
                uploadQueue:
                  type: array
                  items:
                    $ref: "#/components/schemas/UploadQueueEntry"
                servers:
                  type: array
                  items:
                    $ref: "#/components/schemas/Server"
                kad:
                  $ref: "#/components/schemas/Kad"
                network:
                  $ref: "#/components/schemas/NetworkStatus"
                logs:
                  type: array
                  items:
                    $ref: "#/components/schemas/LogEntry"
...
    Transfer:
      type: object
      additionalProperties: false
      required: [hash, name, sizeBytes, state]
      properties:
        hash:
          type: string
        name:
          type: string
        path:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        completedBytes:
          type: integer
          minimum: 0
        state:
          $ref: "#/components/schemas/TransferState"
        priority:
          $ref: "#/components/schemas/TransferPriority"
        categoryId:
          type: integer
          minimum: 0
          maximum: 4294967295
        categoryName:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 1
        sources:
          type: integer
          minimum: 0
        sourcesTransferring:
          type: integer
          minimum: 0
        autoPriority:
          type: boolean
        eta:
          type:
            - integer
            - "null"
        addedAt:
          type:
            - integer
            - "null"
        completedAt:
          type:
            - integer
            - "null"
        partsTotal:
          type: integer
          minimum: 0
        partsObtained:
          type: integer
          minimum: 0
        partsProgressText:
          type: string
        partsAvailable:
          type: integer
          minimum: 0
        stopped:
          type: boolean
        fakeFile:
          $ref: "#/components/schemas/FakeFileAnalysis"
        downloadSpeedKiBps:
          type: number
          minimum: 0
        uploadSpeedKiBps:
          type: number
          minimum: 0
        ed2kLink:
          type: string
    TransferState:
      type: string
      enum: [downloading, paused, queued, checking, completing, completed, error, missingfiles]
    TransferPriority:
      type: string
      enum: [auto, verylow, low, normal, high, veryhigh]
...
    ShutdownRequest:
      type: object
      additionalProperties: false
      required: [confirmShutdown]
      properties:
        confirmShutdown:
          type: boolean
          enum: [true]
...
    SharedFile:
      type: object
      additionalProperties: false
      required: [hash, name, path, directory, sizeBytes, priority, autoUploadPriority, requests, acceptedRequests, transferredBytes, allTimeRequests, allTimeAccepts, allTimeTransferred, partCount, partFile, complete, comment, rating, hasComment, userRating, publishedEd2k, sharedByRule]
      properties:
        hash:
          type: string
        name:
          type: string
        path:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        priority:
          $ref: "#/components/schemas/SharedFilePriority"
        rating:
          type: integer
          minimum: 0
          maximum: 5
        comment:
          type: string
        ed2kLink:
          type: string
        directory:
          type: string
        autoUploadPriority:
          type: boolean
        allTimeRequests:
          type: integer
          minimum: 0
        allTimeAccepts:
          type: integer
          minimum: 0
        allTimeTransferred:
          type: integer
          minimum: 0
        partCount:
          type: integer
          minimum: 0
        partFile:
          type: boolean
        complete:
          type: boolean
        hasComment:
          type: boolean
        userRating:
          type: integer
          minimum: 0
        publishedEd2k:
          type: boolean
        sharedByRule:
          type: boolean
        requests:
          type: integer
          minimum: 0
        acceptedRequests:
          type: integer
          minimum: 0
        transferredBytes:
          type: integer
          minimum: 0
...
    Upload:
      type: object
      additionalProperties: false
      required: [clientId, userName, userHash, clientSoftware, clientMod, uploadState, uploadSpeedKiBps, uploadedBytes, queueSessionUploaded, payloadBuffered, waitTimeMs, waitStartedTick, score, address, port, serverIp, serverPort, lowId, friendSlot, uploading, waitingQueue, requestedFileHash, requestedFileName, requestedFileSizeBytes, requestedPartsObtained, requestedPartsTotal, requestedPartsProgressText]
      properties:
        clientId:
          type: string
        userName:
          type: string
        userHash:
          type:
            - string
            - "null"
        clientSoftware:
          type: string
        clientMod:
          type: string
        uploadState:
          type: string
          enum: [uploading, queued, connecting, banned, idle]
        fileHash:
          type: string
        fileName:
          type: string
        uploadSpeedKiBps:
          type: number
          minimum: 0
        uploadedBytes:
          type: integer
          minimum: 0
        queueSessionUploaded:
          type: integer
          minimum: 0
        payloadBuffered:
          type: integer
          minimum: 0
        waitTimeMs:
          type: integer
          minimum: 0
        waitStartedTick:
          type: integer
          minimum: 0
        score:
          type: integer
          minimum: 0
        address:
          type: string
        port:
          type: integer
          minimum: 0
          maximum: 65535
        serverIp:
          type: string
        serverPort:
          type: integer
          minimum: 0
          maximum: 65535
        lowId:
          type: boolean
        friendSlot:
          type: boolean
        uploading:
          type: boolean
        waitingQueue:
          type: boolean
        requestedFileHash:
          type:
            - string
            - "null"
        requestedFileName:
          type:
            - string
            - "null"
        requestedFileSizeBytes:
          type:
            - integer
            - "null"
          minimum: 0
        requestedPartsObtained:
          type: integer
          minimum: 0
        requestedPartsTotal:
          type: integer
          minimum: 0
        requestedPartsProgressText:
          type: string
        scoreBreakdown:
          $ref: "#/components/schemas/UploadScoreBreakdown"
...
    UploadQueueEntry:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/Upload"
        - type: object
          properties:
            queueRank:
              type: integer
              minimum: 0
...
    Server:
      type: object
      additionalProperties: false
      required: [address, port, name, priority, static, connected, connecting, current, description, dynIp, failedCount, hardFiles, ip, ping, softFiles, version, users, files]
      properties:
        address:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        name:
          type: string
        priority:
          type: string
          enum: [low, normal, high]
        static:
          type: boolean
        connected:
          type: boolean
        connecting:
          type: boolean
        current:
          type: boolean
        description:
          type: string
        dynIp:
          type: string
        failedCount:
          type: integer
          minimum: 0
        hardFiles:
          type: integer
          minimum: 0
        ip:
          type: string
        ping:
          type: integer
          minimum: 0
        softFiles:
          type: integer
          minimum: 0
        version:
          type: string
        users:
          type: integer
          minimum: 0
        files:
          type: integer
          minimum: 0
...
    Kad:
      type: object
      additionalProperties: false
      required: [running, connected, firewalled, bootstrapping, bootstrapProgress, contactCount, lanMode, users, files]
      properties:
        running:
          type: boolean
        connected:
          type: boolean
        firewalled:
          type:
            - boolean
            - "null"
        nodes:
          type: integer
          minimum: 0
        contactCount:
          type:
            - integer
            - "null"
          minimum: 0
        lanMode:
          type: boolean
        users:
          type:
            - integer
            - "null"
          minimum: 0
        files:
          type:
            - integer
            - "null"
          minimum: 0
        bootstrapping:
          type: boolean
        bootstrapProgress:
          type: integer
          minimum: 0
        indexedSources:
          type: integer
          minimum: 0
        indexedKeywords:
          type: integer
          minimum: 0
        operationQueued:
          type: boolean
        alreadyRunning:
          type: boolean
        blockedByVpnGuard:
          type: boolean
        network:
          $ref: "#/components/schemas/NetworkStatus"
...
    LogEntry:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        timestamp:
          oneOf:
            - type: string
              format: date-time
            - type: integer
        level:
          type: string
        debug:
          type: boolean
        message:
          type: string
...